webvh-e2e: rename e2e_constants module to constants
Made-with: Cursor
